+++ HISTORY.rst
@@ -50,3 +50,4 @@ History
 ------------------
 * Allowing to provide `correlation_id` value when the client is created
 * Caching tokens by `grant_client_id` and `token_url` to avoid calling identity in case credentials are cached
+* PageBrowser keeping a response cache to avoid duplicated requests